requirements.txt
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
phonemizer==3.3.0
polars==1.6.0
python_Levenshtein==0.25.1
tqdm==4.66.5